+++ R/seasonal.R
@@ -18,10 +18,7 @@
 #' @export
 #' @import dplyr
 #' @examples
-#' org <- AirPassengers %>%
-#'   tibble(time = lubridate::date_decimal(as.numeric(time(.))), values = .)
-#'
-#' y <- sa_series(x = org$values, time = org$time)
+#' y <- sa_series(x = AirPassengers, time = zoo::as.Date(AirPassengers))
 
 seasonal_adj <- function(x, time, series = "sa", outlier.types = "ao", ...){