+++ vignettes/Introduction.R
@@ -8,16 +8,16 @@ library(forcats)
 ## ----data----------------------------------------------------------------
 
 dat_ku <- pxweb::get_pxweb_data(
-  url = "http://pxnet2.stat.fi/PXWeb/api/v1/fi/StatFin/vrm/tyokay/010_tyokay_tau_101.px",
-  dims = list(
-    Alue = c('*'),
-    "Pääasiallinen toiminta" = c('11'),  # Työlliset
-    Sukupuoli = c('S'),
-    "Ikä" = c('SSS'),
-    Vuosi = c('*')),
-  clean = TRUE) %>% 
+    url = "http://pxnet2.stat.fi/PXWeb/api/v1/fi/StatFin/vrm/tyokay/statfin_tyokay_pxt_001.px",
+    dims = list(Alue = c('*'),
+               "Pääasiallinen toiminta" = c('11'),
+               Sukupuoli = c('S'),
+               "Ikä" = c('SSS'),
+               Vuosi = c('*')),
+    clean = TRUE) %>% 
   clean_names() %>% 
   clean_times()
+
 
 
 ## ----aggregate-----------------------------------------------------------